Update libft and push_swap code. String arguments

+++ libft/ft_atoi.c
@@ -6,18 +6,18 @@
 /*   By: issierra <issierra@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/19 13:50:11 by issierra          #+#    #+#             */
-/*   Updated: 2023/09/29 11:14:26 by issierra         ###   ########.fr       */
+/*   Updated: 2023/12/23 11:47:51 by issierra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 // The atoi() function converts the initial portion of the string pointed to 
 // by str to int representation.
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
 	int	i;
 	int	sign;
-	int	res;
+	long int	res;
 
 	i = 0;
 	sign = 1;
